Refactor and separate ScanningDevice.cs to Imaging namespace
